Thando Tinto - Fixed Payments Repository, added Ticket Repository. Deleted all code you made that was not your duty to do because they were my classes. NB: Stop touching my things without letting me know.
