+++ src/main/java/za/ac/cput/Factory/payment/PaymentFactory.java
@@ -1,6 +1,6 @@
-package za.ac.cput.Factory;
+package za.ac.cput.Factory.payment;
 
-import za.ac.cput.Domain.Payment;
+import za.ac.cput.Domain.payment.Payment;
 import za.ac.cput.Util.Helper;
 
 import java.time.LocalDate;
