+++ src/main/java/za/ac/cput/Repository/TicketRepository.java
@@ -2,9 +2,8 @@ package za.ac.cput.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import za.ac.cput.Domain.payment.Payment;
-
+import za.ac.cput.Domain.payment.Ticket;
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment,Integer> {
+public interface TicketRepository extends JpaRepository<Ticket,Integer> {
 }
