+++ src/test/java/za/ac/cput/Factory/TicketFactoryTest.java
@@ -2,7 +2,8 @@ package za.ac.cput.Factory;
 
 
 import org.junit.jupiter.api.*;
-import za.ac.cput.Domain.Ticket;
+import za.ac.cput.Domain.payment.Ticket;
+import za.ac.cput.Factory.payment.TicketFactory;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
